Support auto-serialization (a feature in new versions of Mrs)
Note that there is a current bug wherein particles are mutated.  This bug
leads to a race condition that makes results in the serial implementation
a bit off.  Although this is probably technically a bug in the PSO code,
it's probably better to work around this in Mrs rather than to make parallel
PSO less efficient.